+++ packages/stack/src/lib/stack-app.ts
@@ -736,7 +736,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       clientMetadata: crud.client_metadata,
       clientReadOnlyMetadata: crud.client_read_only_metadata,
       async inviteUser(options: { email: string }) {
-        return await app._interface.sendTeamInvitation({
+        await app._interface.sendTeamInvitation({
           teamId: crud.id,
           email: options.email,
           session: app._getSession(),
@@ -1916,7 +1916,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
         await app._serverTeamMemberProfilesCache.refresh([crud.id]);
       },
       async inviteUser(options: { email: string }) {
-        return await app._interface.sendTeamInvitation({
+        await app._interface.sendTeamInvitation({
           teamId: crud.id,
           email: options.email,
           session: null,
@@ -2727,6 +2727,7 @@ type ServerUserCreateOptions = {
   primaryEmail?: string,
   primaryEmailAuthEnabled?: boolean,
   password?: string,
+  otpAuthEnabled?: boolean,
   displayName?: string,
   primaryEmailVerified?: boolean,
 }
@@ -2734,6 +2735,7 @@ function serverUserCreateOptionsToCrud(options: ServerUserCreateOptions): UsersC
   return {
     primary_email: options.primaryEmail,
     password: options.password,
+    otp_auth_enabled: options.otpAuthEnabled,
     primary_email_auth_enabled: options.primaryEmailAuthEnabled,
     display_name: options.displayName,
     primary_email_verified: options.primaryEmailVerified,
@@ -3000,7 +3002,7 @@ export type Team = {
   profileImageUrl: string | null,
   clientMetadata: any,
   clientReadOnlyMetadata: any,
-  inviteUser(options: { email: string }): Promise<Result<undefined, KnownErrors["TeamPermissionRequired"]>>,
+  inviteUser(options: { email: string }): Promise<void>,
   listUsers(): Promise<TeamUser[]>,
   useUsers(): TeamUser[],
   update(update: TeamUpdateOptions): Promise<void>,
@@ -3046,7 +3048,7 @@ export type ServerTeam = {
   update(update: ServerTeamUpdateOptions): Promise<void>,
   delete(): Promise<void>,
   addUser(userId: string): Promise<void>,
-  inviteUser(options: { email: string }): Promise<Result<undefined, KnownErrors["TeamPermissionRequired"]>>,
+  inviteUser(options: { email: string }): Promise<void>,
   removeUser(userId: string): Promise<void>,
 } & Team;
 
@@ -3056,7 +3058,6 @@ function serverTeamCreateOptionsToCrud(options: ServerTeamCreateOptions): TeamsC
 }
 
 export type ServerTeamUpdateOptions = TeamUpdateOptions & {
-
   clientReadOnlyMetadata?: ReadonlyJson,
   serverMetadata?: ReadonlyJson,
 };