+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @stackframe/oauth-mock-server
 
+## 2.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated account settings page
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.12
+
 ## 2.6.11
 
 ### Patch Changes