+++ examples/demo/CHANGELOG.md
@@ -1,4 +1,23 @@
-# demo-app
+# @stackframe/example-demo-app
+
+## 2.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.13
+  - @stackframe/stack@2.6.13
+  - @stackframe/stack-ui@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated account settings page
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.12
+  - @stackframe/stack-ui@2.6.12
+  - @stackframe/stack@2.6.12
 
 ## 2.6.11